+++ bonus/inc/cub3D.h
@@ -307,6 +307,7 @@ int release_key(int keycode, t_game *game);
 int has_wall_for_raycast(t_game *game, double x, double y);
 int has_blocking_tile(t_game *game, double x, double y);
 int collide_diagonal(t_game *game, double to_x, double to_y);
+
 void    calculate_next_step(t_game *game, int move, int side_move);
 void    bound_angle(double *angle);
 void    player_movement(t_game *game);
